Refactor PluginIntegrationPanel component
