+++ projects/packages/forms/src/blocks/contact-form/components/shared/plugin-integration-panel/styles.css
@@ -1,0 +1,22 @@
+.jetpack-plugin-integration__status {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+}
+
+.jetpack-plugin-integration__content {
+	color: rgba(38, 46, 57, 0.7);
+}
+
+.jetpack-contact-form__integration-panel-content > *:not(:last-child) {
+	margin-bottom: 16px;
+}
+
+.jetpack-plugin-integration__spinner-icon {
+	animation: rotation 2s infinite linear;
+}
+
+.jetpack-contact-form__akismet-buttons {
+	display: flex;
+	gap: 8px;
+}
